requirements.txt
serial
adafruit-circuitpython-bmp280
RPi.GPIO
smbus
pyserial
